Remove NNSContract.RegisterTable.cs
